test: extract duplicated subprocess crash harness into tests/common (#327)
wal_test.rs and migration_matrix_test.rs carried byte-identical copies of
the crash-simulation harness (CRASH_CHILD_SPAWN_LOCK, CrashTx,
run_crashing_child, crash_child_entrypoint). Moved into tests/common/mod.rs
so there's a single source of truth; both files now `mod common;`.

Each integration test binary still compiles tests/common/mod.rs
separately, so the child-process dispatch name changes from
"crash_child_entrypoint" to "common::crash_child_entrypoint". The
#[allow(dead_code)] on CrashTx::Explicit moves to its one definition site,
since migration_matrix_test's binary still never constructs that variant.
